Stubs ready for code review.
