+++ SubZero/src/util/filter/Filter.h
@@ -17,41 +17,86 @@ class Filter {
 	 * Filter is a virtual class that is further specified into specific
 	 * filters accessible by FilterFactory. In general filters will have
 	 * a filter() function that will filter input according to the
-	 * algorithm it contains and the parameters it holds.
+	 * algorithm it contains and the parameters it holds thus overriding
+	 * the virtual filter() of this parent class.
 	 */
 
 public:
 	Filter();
 	virtual ~Filter();
 
-	// The filter function takes a pointer and applies the algorithm of the
-	// specific filter to the content of the pointer. The content of the
-	// pointer is altered and 0 is returned for successful filter, 1 for no
-	// action performed due to incorrect type, 2 for other errors.
-	virtual int filter(ImgData* camData) {
-		/*cv::Mat temp = camData->img;
-		//do filtration on temp
-		free(camData);
-		camData = &(new ImgData(ID, temp)); // operator overload?
-		return 1;*/
-	}
-	// Overload for FPGA filtering
-	virtual int filter(FPGAData* fpgaData) {
+	/* ==========================================================================
+	 * METHODS
+	 * Which are inherit to all filters
+	 */
+
+
+	/* The filter function takes a pointer and applies the algorithm of the
+	 * specific filter to the content of the pointer. The content of the
+	 * pointer is altered.
+	 *
+	 * @param ImgData to be processed
+	 * @return 0 for success 1 for incorrect type.
+	 */
+	virtual int filter(ImgData* &imgData,) {
 		return 1;
 	}
 
-	// Get the ID of the specific filter instance
+	/*
+	 * Overload for FPGA filtering
+	 *
+	 * @param FPGAData to be processed
+	 * @return 0 for success 1 for incorrect type.
+	 */
+	virtual int filter(FPGAData* &fpgaData) {
+		return 1;
+	}
+
+	/*
+	 * Get the ID of the specific filter instance
+	 *
+	 * @return string of ID.
+	 */
 	std::string getID();
 
 	/*
 	 * Set ID of a filter.
 	 *
 	 * @param string ID
+	 * return 0 for success, 1 for fail.
 	 */
-	void setID(std::string ID);
+	int setID(std::string ID);
 
-private:
+	/*
+	 * Get the msg of the filter object.
+	 *
+	 * @return string msg.
+	 */
+	std::string getMsg();
+
+	/*
+	 * Set a msg to a filter object.
+	 *
+	 * @param string msg to be propagated.
+	 * @return 0 for normal operation, 1 for warning of previous content.
+	 */
+	int setMsg();
+
+
+protected:
+	/* ==========================================================================
+	 * CLASS VARIABLES
+	 * Theses variables are accessible by children of filters.
+	 */
+
+	/*
+	 * Identifier for the filter.
+	 */
 	std::string ID;
+
+	/*
+	 * Message propagated by the filter.
+	 */
 	std::string msg;
 };
